+++ WEB/frontend/src/pages/Board.js
